updated JASMIN WVCCI_INST scripts

+++ src/main/bin/cems_shared/wvcci-inst/bin/wvcci-l2-idepix-modis-snap.sh
@@ -8,7 +8,6 @@ sensor=$5
 year=$6
 month=$7
 wvcciRootDir=$8
-snapRootDir=$9
 
 if [ "$sensor" == "MERIS" ]
 then
@@ -28,14 +27,10 @@ then
     mkdir -p $idepixL2Dir
 fi
 
-if [ "$sensor" == "MERIS" ]
+if [ "$sensor" == "MODIS_TERRA" ]
 then
-    echo "time $snapRootDir/bin/gpt Snap.Idepix.Meris -e -SsourceProduct=$l1bPath -PreflBandsToCopy=reflectance_13,reflectance_14,reflectance_15 -f NetCDF4-IDEPIX-MERIS -t $idepixFile"
-    time $snapRootDir/bin/gpt Snap.Idepix.Meris -e -SsourceProduct=$l1bPath -PreflBandsToCopy=reflectance_13,reflectance_14,reflectance_15 -f NetCDF4-IDEPIX-MERIS -t $idepixFile
-elif [ "$sensor" == "MODIS_TERRA" ]
-then
-    echo "time $snapRootDir/bin/gpt Snap.Idepix.Modis -e -SsourceProduct=$l1bPath -SmodisWaterMask=$landMaskPath -PreflBandsToCopy=EV_250_Aggr1km_RefSB_2,EV_500_Aggr1km_RefSB_5,EV_1KM_RefSB_17,EV_1KM_RefSB_18,EV_1KM_RefSB_19 -f NetCDF4-IDEPIX-MODIS -t $idepixFile"
-    time $snapRootDir/bin/gpt Snap.Idepix.Modis -e -SsourceProduct=$l1bPath -SmodisWaterMask=$landMaskPath -PreflBandsToCopy=EV_250_Aggr1km_RefSB_2,EV_500_Aggr1km_RefSB_5,EV_1KM_RefSB_17,EV_1KM_RefSB_18,EV_1KM_RefSB_19 -f NetCDF4-IDEPIX-MODIS -t $idepixFile
+    echo "time ${SNAP_HOME}/bin/gpt Idepix.Modis -e -SsourceProduct=$l1bPath -SmodisWaterMask=$landMaskPath -PreflBandsToCopy=EV_250_Aggr1km_RefSB_2,EV_500_Aggr1km_RefSB_5,EV_1KM_RefSB_17,EV_1KM_RefSB_18,EV_1KM_RefSB_19 -f NetCDF4-IDEPIX-MODIS -t $idepixFile"
+    time ${SNAP_HOME}/bin/gpt Idepix.Modis -e -SsourceProduct=$l1bPath -SmodisWaterMask=$landMaskPath -PreflBandsToCopy=EV_250_Aggr1km_RefSB_2,EV_500_Aggr1km_RefSB_5,EV_1KM_RefSB_17,EV_1KM_RefSB_18,EV_1KM_RefSB_19 -f NetCDF4-IDEPIX-MODIS -t $idepixFile
 else
     echo "Invalid sensor $sensor - no Idepix processing started."
 fi
@@ -48,9 +43,13 @@ then
     echo "Idepix product created."
     echo "Status: $status"
 else
-    echo "Idepix product NOT successfully created."
-    echo "rm -f $idepixFile"
-    rm -f $idepixFile   # delete corrupt file
+    echo "Idepix product NOT successfully created (corrupt or not a Day product)."
+    echo "Status: $status"
+    if [ -e "$idepixFile" ]
+    then
+      echo "rm -f $idepixFile"
+      rm -f $idepixFile   # delete corrupt file
+    fi
 fi
 
 echo `date`
